+++ source/extensions/filters/common/rbac/matchers.cc
@@ -8,13 +8,17 @@ namespace Filters {
 namespace Common {
 namespace RBAC {
 
-MatcherConstSharedPtr Matcher::create(const envoy::config::rbac::v2alpha::Permission& permission) {
+MatcherConstSharedPtr Matcher::create(const envoy::config::rbac::v2alpha::Permission& permission,
+                                      bool disable_http_rules) {
   switch (permission.rule_case()) {
   case envoy::config::rbac::v2alpha::Permission::RuleCase::kAndRules:
-    return std::make_shared<const AndMatcher>(permission.and_rules());
+    return std::make_shared<const AndMatcher>(permission.and_rules(), disable_http_rules);
   case envoy::config::rbac::v2alpha::Permission::RuleCase::kOrRules:
-    return std::make_shared<const OrMatcher>(permission.or_rules());
+    return std::make_shared<const OrMatcher>(permission.or_rules(), disable_http_rules);
   case envoy::config::rbac::v2alpha::Permission::RuleCase::kHeader:
+    if (disable_http_rules) {
+      return std::make_shared<const AlwaysMatcher>();
+    }
     return std::make_shared<const HeaderMatcher>(permission.header());
   case envoy::config::rbac::v2alpha::Permission::RuleCase::kDestinationIp:
     return std::make_shared<const IPMatcher>(permission.destination_ip(), true);
@@ -23,46 +27,58 @@ MatcherConstSharedPtr Matcher::create(const envoy::config::rbac::v2alpha::Permis
   case envoy::config::rbac::v2alpha::Permission::RuleCase::kAny:
     return std::make_shared<const AlwaysMatcher>();
   case envoy::config::rbac::v2alpha::Permission::RuleCase::kMetadata:
+    if (disable_http_rules) {
+      return std::make_shared<const AlwaysMatcher>();
+    }
     return std::make_shared<const MetadataMatcher>(permission.metadata());
   case envoy::config::rbac::v2alpha::Permission::RuleCase::kNotRule:
-    return std::make_shared<const NotMatcher>(permission.not_rule());
+    return std::make_shared<const NotMatcher>(permission.not_rule(), disable_http_rules);
   default:
     NOT_REACHED_GCOVR_EXCL_LINE;
   }
 }
 
-MatcherConstSharedPtr Matcher::create(const envoy::config::rbac::v2alpha::Principal& principal) {
+MatcherConstSharedPtr Matcher::create(const envoy::config::rbac::v2alpha::Principal& principal,
+                                      bool disable_http_rules) {
   switch (principal.identifier_case()) {
   case envoy::config::rbac::v2alpha::Principal::IdentifierCase::kAndIds:
-    return std::make_shared<const AndMatcher>(principal.and_ids());
+    return std::make_shared<const AndMatcher>(principal.and_ids(), disable_http_rules);
   case envoy::config::rbac::v2alpha::Principal::IdentifierCase::kOrIds:
-    return std::make_shared<const OrMatcher>(principal.or_ids());
+    return std::make_shared<const OrMatcher>(principal.or_ids(), disable_http_rules);
   case envoy::config::rbac::v2alpha::Principal::IdentifierCase::kAuthenticated:
     return std::make_shared<const AuthenticatedMatcher>(principal.authenticated());
   case envoy::config::rbac::v2alpha::Principal::IdentifierCase::kSourceIp:
     return std::make_shared<const IPMatcher>(principal.source_ip(), false);
   case envoy::config::rbac::v2alpha::Principal::IdentifierCase::kHeader:
+    if (disable_http_rules) {
+      return std::make_shared<const AlwaysMatcher>();
+    }
     return std::make_shared<const HeaderMatcher>(principal.header());
   case envoy::config::rbac::v2alpha::Principal::IdentifierCase::kAny:
     return std::make_shared<const AlwaysMatcher>();
   case envoy::config::rbac::v2alpha::Principal::IdentifierCase::kMetadata:
+    if (disable_http_rules) {
+      return std::make_shared<const AlwaysMatcher>();
+    }
     return std::make_shared<const MetadataMatcher>(principal.metadata());
   case envoy::config::rbac::v2alpha::Principal::IdentifierCase::kNotId:
-    return std::make_shared<const NotMatcher>(principal.not_id());
+    return std::make_shared<const NotMatcher>(principal.not_id(), disable_http_rules);
   default:
     NOT_REACHED_GCOVR_EXCL_LINE;
   }
 }
 
-AndMatcher::AndMatcher(const envoy::config::rbac::v2alpha::Permission_Set& set) {
+AndMatcher::AndMatcher(const envoy::config::rbac::v2alpha::Permission_Set& set,
+                       bool disable_http_rules) {
   for (const auto& rule : set.rules()) {
-    matchers_.push_back(Matcher::create(rule));
+    matchers_.push_back(Matcher::create(rule, disable_http_rules));
   }
 }
 
-AndMatcher::AndMatcher(const envoy::config::rbac::v2alpha::Principal_Set& set) {
+AndMatcher::AndMatcher(const envoy::config::rbac::v2alpha::Principal_Set& set,
+                       bool disable_http_rules) {
   for (const auto& id : set.ids()) {
-    matchers_.push_back(Matcher::create(id));
+    matchers_.push_back(Matcher::create(id, disable_http_rules));
   }
 }
 
@@ -79,16 +95,18 @@ bool AndMatcher::matches(const Network::Connection& connection,
 }
 
 OrMatcher::OrMatcher(
-    const Protobuf::RepeatedPtrField<::envoy::config::rbac::v2alpha::Permission>& rules) {
+    const Protobuf::RepeatedPtrField<::envoy::config::rbac::v2alpha::Permission>& rules,
+    bool disable_http_rules) {
   for (const auto& rule : rules) {
-    matchers_.push_back(Matcher::create(rule));
+    matchers_.push_back(Matcher::create(rule, disable_http_rules));
   }
 }
 
 OrMatcher::OrMatcher(
-    const Protobuf::RepeatedPtrField<::envoy::config::rbac::v2alpha::Principal>& ids) {
+    const Protobuf::RepeatedPtrField<::envoy::config::rbac::v2alpha::Principal>& ids,
+    bool disable_http_rules) {
   for (const auto& id : ids) {
-    matchers_.push_back(Matcher::create(id));
+    matchers_.push_back(Matcher::create(id, disable_http_rules));
   }
 }