+++ source/extensions/filters/http/rbac/rbac_filter.cc
@@ -16,60 +16,6 @@ static const std::string resp_code_403 = "403";
 static const std::string shadow_policy_id_field = "shadow_effective_policyID";
 static const std::string shadow_resp_code_field = "shadow_response_code";
 
-RoleBasedAccessControlFilterConfig::RoleBasedAccessControlFilterConfig(
-    const envoy::config::filter::http::rbac::v2::RBAC& proto_config,
-    const std::string& stats_prefix, Stats::Scope& scope)
-    : stats_(RoleBasedAccessControlFilter::generateStats(stats_prefix, scope)),
-      engine_(proto_config.has_rules()
-                  ? absl::make_optional<Filters::Common::RBAC::RoleBasedAccessControlEngineImpl>(
-                        proto_config.rules())
-                  : absl::nullopt),
-      shadow_engine_(
-          proto_config.has_shadow_rules()
-              ? absl::make_optional<Filters::Common::RBAC::RoleBasedAccessControlEngineImpl>(
-                    proto_config.shadow_rules())
-              : absl::nullopt) {}
-
-RoleBasedAccessControlFilterStats
-RoleBasedAccessControlFilter::generateStats(const std::string& prefix, Stats::Scope& scope) {
-  const std::string final_prefix = prefix + "rbac.";
-  return {ALL_RBAC_FILTER_STATS(POOL_COUNTER_PREFIX(scope, final_prefix))};
-}
-
-const absl::optional<Filters::Common::RBAC::RoleBasedAccessControlEngineImpl>&
-RoleBasedAccessControlFilterConfig::engine(const Router::RouteConstSharedPtr route,
-                                           EnforcementMode mode) const {
-  if (!route || !route->routeEntry()) {
-    return engine(mode);
-  }
-
-  const std::string& name = HttpFilterNames::get().Rbac;
-  const auto* entry = route->routeEntry();
-
-  const auto* route_local =
-      entry->perFilterConfigTyped<RoleBasedAccessControlRouteSpecificFilterConfig>(name)
-          ?: entry->virtualHost()
-                 .perFilterConfigTyped<RoleBasedAccessControlRouteSpecificFilterConfig>(name);
-
-  if (route_local) {
-    return route_local->engine(mode);
-  }
-
-  return engine(mode);
-}
-
-RoleBasedAccessControlRouteSpecificFilterConfig::RoleBasedAccessControlRouteSpecificFilterConfig(
-    const envoy::config::filter::http::rbac::v2::RBACPerRoute& per_route_config)
-    : engine_(per_route_config.rbac().has_rules()
-                  ? absl::make_optional<Filters::Common::RBAC::RoleBasedAccessControlEngineImpl>(
-                        per_route_config.rbac().rules())
-                  : absl::nullopt),
-      shadow_engine_(
-          per_route_config.rbac().has_shadow_rules()
-              ? absl::make_optional<Filters::Common::RBAC::RoleBasedAccessControlEngineImpl>(
-                    per_route_config.rbac().shadow_rules())
-              : absl::nullopt) {}
-
 Http::FilterHeadersStatus RoleBasedAccessControlFilter::decodeHeaders(Http::HeaderMap& headers,
                                                                       bool) {
   ENVOY_LOG(
@@ -84,9 +30,12 @@ Http::FilterHeadersStatus RoleBasedAccessControlFilter::decodeHeaders(Http::Head
                 callbacks_->connection()->ssl()->subjectPeerCertificate()
           : "none",
       headers, callbacks_->requestInfo().dynamicMetadata().DebugString());
+
   std::string effective_policy_id;
-  const absl::optional<Filters::Common::RBAC::RoleBasedAccessControlEngineImpl>& shadow_engine =
-      config_->engine(callbacks_->route(), EnforcementMode::Shadow);
+  const std::string& filter_name = HttpFilterNames::get().Rbac;
+  const auto& shadow_engine = config_->engine(callbacks_->route(), filter_name,
+                                              Filters::Common::RBAC::EnforcementMode::Shadow);
+
   if (shadow_engine.has_value()) {
     std::string shadow_resp_code = resp_code_200;
     if (shadow_engine->allowed(*callbacks_->connection(), headers,
@@ -119,8 +68,8 @@ Http::FilterHeadersStatus RoleBasedAccessControlFilter::decodeHeaders(Http::Head
     }
   }
 
-  const absl::optional<Filters::Common::RBAC::RoleBasedAccessControlEngineImpl>& engine =
-      config_->engine(callbacks_->route(), EnforcementMode::Enforced);
+  const auto& engine = config_->engine(callbacks_->route(), filter_name,
+                                       Filters::Common::RBAC::EnforcementMode::Enforced);
   if (engine.has_value()) {
     if (engine->allowed(*callbacks_->connection(), headers,
                         callbacks_->requestInfo().dynamicMetadata(), nullptr)) {